refactor(cmd): moving files around to follow closer to best practices outlined in helm
