+++ cmd/eval_test.go
@@ -1,4 +1,4 @@
-package eval
+package cmd
 
 import (
 	"testing"
