+++ cmd/root.go
@@ -1,0 +1,17 @@
+package cmd
+
+import (
+	"github.com/spf13/cobra"
+)
+
+var rootCmd = &cobra.Command{
+	Use:   "combo",
+	Short: "Create combinations of Kubernetes manifests",
+}
+
+// Execute executes the root command.
+func Execute() error {
+	rootCmd.CompletionOptions.DisableDefaultCmd = true
+
+	return rootCmd.Execute()
+}
