+++ cmd/root_test.go
@@ -1,0 +1,9 @@
+package cmd
+
+import "testing"
+
+func TestRoot(t *testing.T) {
+	t.Run("Execute", func(t *testing.T) {
+		Execute()
+	})
+}
