--- internal/cli/root/root.go
@@ -1,22 +1,0 @@
-package root
-
-import (
-	"github.com/operator-framework/combo/internal/cli/eval"
-	"github.com/spf13/cobra"
-)
-
-func NewCmd() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "combo",
-		Short: "Create combinations of Kubernetes manifests",
-	}
-
-	// Disable the default auto-completion command
-	cmd.CompletionOptions.DisableDefaultCmd = true
-
-	cmd.AddCommand(
-		eval.NewCommand(),
-	)
-
-	return cmd
-}
